pyproject.toml
[project]
name = "somethingelse"
version = "1.0.0"
description = "Aircraft tracking UI using ADSB and Weather data"
requires-python = ">=3.12"
dependencies = [
    "pywebview>=6.1",
]

[project.scripts]
somethingelse = "somethingelse.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/somethingelse"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
